+++ packages/core/src/blade/input-binding/api/input-binding-test.ts
@@ -3,12 +3,12 @@ import {describe, it} from 'mocha';
 
 import {Binding} from '../../../common/binding/binding';
 import {BindingTarget} from '../../../common/binding/target';
+import {BindingValue} from '../../../common/binding/value';
 import {
 	createNumberFormatter,
 	numberFromUnknown,
 	parseNumber,
 } from '../../../common/converter/number';
-import {BindingValue} from '../../../common/model/binding-value';
 import {ValueMap} from '../../../common/model/value-map';
 import {createValue} from '../../../common/model/values';
 import {ViewProps} from '../../../common/model/view-props';
