+++ packages/core/src/input-binding/plugin.ts
@@ -4,8 +4,8 @@ import {LabeledValueController} from '../blade/label/controller/value-label';
 import {LabelPropsObject} from '../blade/label/view/label';
 import {Binding, BindingReader, BindingWriter} from '../common/binding/binding';
 import {BindingTarget} from '../common/binding/target';
+import {BindingValue} from '../common/binding/value';
 import {Constraint} from '../common/constraint/constraint';
-import {BindingValue} from '../common/model/binding-value';
 import {Value} from '../common/model/value';
 import {ValueMap} from '../common/model/value-map';
 import {createValue} from '../common/model/values';
