+++ packages/core/src/input-binding/point-4d/plugin-test.ts
@@ -2,9 +2,9 @@ import * as assert from 'assert';
 import {describe, it} from 'mocha';
 
 import {BindingTarget} from '../../common/binding/target';
+import {BindingValue} from '../../common/binding/value';
 import {findConstraint} from '../../common/constraint/composite';
 import {StepConstraint} from '../../common/constraint/step';
-import {BindingValue} from '../../common/model/binding-value';
 import {ComplexValue} from '../../common/model/complex-value';
 import {getBoundValue} from '../../common/model/test-util';
 import {createTestWindow} from '../../misc/dom-test-util';
